pyproject.toml
[tool.poetry]
name = "atlas-tracing"
version = "0.1.0"
description = ""
authors = ["Guilherme Scaranse <guilherme@joinsaturn.com>"]

[tool.poetry.dependencies]
python = "^3.7"
opentelemetry-api = "^0.16b1"
opentelemetry-sdk = "^0.16b1"
opentelemetry-instrumentation-asgi = "^0.16b1"
opentelemetry-instrumentation-sqlalchemy = "^0.16b1"
opentelemetry-instrumentation-requests = "^0.16b1"
opentelemetry-instrumentation-redis = "^0.16b1"
opentelemetry-instrumentation-fastapi = "^0.16b1"
opentelemetry-instrumentation-botocore = "^0.16b1"
opentelemetry-instrumentation-boto = "^0.16b1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
